FEAT: Add color picker button
Added color picker button on CreateScreen. If you tap on the button, a
dialog will comes up and show you the color that you can create and
choose.

The next thing to do is to make it keep showing the color even after I
closed the dialog. And on the next step, the users may can see the color
of the clothes on the background has changed.