FEAT: Implement color changing by picking color
On CreateScreen, if you tap on the 'color' button, a dialog will comes
up and it'll help you to pick a color. If it's done, you can touch the
outside of the dialog then it'll be closed. And then, you'll see that
your color is applied to the background color!