FEAT: Set the backgroudn to have the initial color
When the button is cliked after the color was set, there was a problem
that it keep refreshing its color to white. It has initialColor
parameter but it also didn't work even the value was in view model. So
just added a variable in the view model that save the color when
dismissing the alert dialog and set that as the initialColor. And
thankfully now it works well as I wanted!