FEAT: Fix bug and show clothes category.
First, I changed the api cuz the one that I used before was way too
inefficient for minimizing the api call and too slow. And then faced
some a lot of bugs related to api and now when you tap on the
SearchScreen, you can see that the category list is showing up well!