FEAT: Show item details when image is clicked
If the image of an item is clicked, it navigates to a new page and show
the item details with more images. At first, there were some problems
with showing the image but now it works fine.

After this, need to implement the search function if possible and show
the corresponding result of the keyword. Plus, need some styling on the
page that I worked on today for better readability and style.