FEAT: Implement search bar function
Fully implemented the search bar function with loading the search
histories, deleting histories, and pasting the saved history right into
the search bar. Used room database to implement this feature. And
nicely the search history list gets changed as soon as the histories got
deleted lively thx to the LiveData class.

Now need to connect to the api so that the users can type any keyword in
the searchbar, and find some corresponding items on the screen.