FIX: Fix LazyRow to shows up properly
Fixed the bug that the LazyRow didn't show up properly on the DetailScreen.
Now, it's completely fixed and the LazyRow with the "You might also
like" section is showing up very well.

Plus, if you tap on the each item of the list, it navigates to the product's
detail screen and again show the "You might also like" section" at the
bottom. And when navigating back, it goes back step by step as it
navigated into.