FEAT: Implement wish list function on MyPageScreen
Implemented additional new database for saving the items in the wish
list. And now every item's detail screen has bookmark icon. If it's
already in wish list, it is not filled but if not, it's full.

If there's nothing in the wish list and try go to the MyPageScreen.
It'll show the texts saying that there's nothing in the wish list and
recommending the users to find the items that they like and to add the
items in the wishlist. And if there're something in the wish list, it'll
show items that have been saved in the database.