REFACTOR: Change layout on CreateScreen
Changed the location of the clothes(sweater and pants) on CreateScreen.
These were locating exactly the center of their places so when running
this app, these two were being apart, which looks kinda awkward. So
changed the vertical arrangement style of these two so that it looks
like they are attached.